Add reverse proxy option and fix proxy mode
"-allow-proxy" would not set IP correctly (confusing sockaddr and in_addr
types). This commit rewrites that code and adds a new feature.

Reverse proxy mode is for use when hefur is running behind something
like nginx, which obscures the request ip, but can send it in a HTTP
header. In the example nginx configuration sample:
	location = /announce {
		proxy_set_header Host $http_host;
		proxy_set_header X-Forwarded-Proto $scheme;
		proxy_set_header X-Real-IP $remote_addr;
		proxy_set_header X-Forwarded-For $proxy_add_x_forwarded_for;
		proxy_redirect off;
		proxy_pass http://127.0.0.1:6969/announce;
	}
Announces would be passed along to hefur with the request IP in the
header X-Real-IP. Running hefur with
	-reverse-proxy-header X-Real-IP
would correctly set IPs.

Now, on announce, the IP comes from:
- if allow-proxy: ip query string
- else if reverse-proxy: HTTP header specified
- else: request ip